Brand New Tutorial Scene
A new tutorial scene ("Intro") that contains 3 finished puzzles and 1 in
the works. Designed so that the player only needs to shoot blue.

Besides the obvious additions, I have gone through and renamed all of
the entries in the hierarchy to be easier to find/understand by name.

Issues/Necessary Fixes:
--  The already-red items in the level can still be made blue by the
player; need to find a way to lock these in the red polarity.
-- The fourth room's puzzle does not currently work due to the right
wall having a screwy magnetic field (raises blocks when should move
left-right).
-- The hand prefab is missing.