Attempted to Fix Room 5 Errors (*UNSOLVED*)
Current issues with Room 5:
-- The pads sliding along the ground are getting caught on nonexistent
colliders between GameObjects. The floor is made of three cubes of
identical widths, but the pads often get caught on the line separating
the cubes.
-- To try and fix the above problem, I adjusted the boxcolliders of the
sliding pads; this did not help, but did cause the pads to get stuck
inside walls less often.

General Issues:
-- Cubic magnetic fields are uneven; more force is applied based on
distance from the center (an item at the far-right end of the room moves
differently than one near the center-right when the wall is magnetised).