MAJOR EDITS PLEASE READ
I have adjusted the gravity of the game!! This was to make the cubes
fall more realistically. As a result, the magnetic force of some cubes
has been increased, and some walls have been lowered. I believe that
this affects all scenes.

The new Tut_Rev scene will be our tutorial. It needs doors that open on
a trigger. Along with this, the wall magnets in rooms 4 and 5 need to be
replaced with smaller wall pads to make the growth of magnets more
gradual.