Lots of Changes!
Changelog:
-- Levels 1 and 2 are now titled "Repelling Tutorial" and "Attracting
Tutorial".
-- Level 2 has been started! Please note that only the first two rooms
are complete; the remaining three are copies from Level 1. All puzzles
are solvable by only shooting blue.
-- Buttons are less eye-catching.
-- Made Level 1 Room 5 simpler.

To-Do List:
-- Button model.
-- Magnetic field fixes.
-- Level 1 Room 5 needs a reason to not drop into the pit. I'm thinking
lasers stretching from one platform to the other a little below the lip.
These lasers should be visible and seem dangerous, but thin enough to
see the pieces below and be able to solve the puzzle. There should also
be a "reset" mechanic in place that resets the puzzle if the person
falls in and dies.
-- Need to have a way for items to be magnetized but nor changable by
the player. We need to be careful in doing this as players may get
confused why they can or can't change certain objects.
-- Make crosshair get thicker/visible cue that an object being pointed
at is magnetizable.