Made Certain Blocks Unchangable!
The player's understanding of what can be changed and what can't will
come with whatever textures we use, but as far as coding goes I made a
simple script that deems an object unchangable! This also means that the
crosshair does not change on this object.

Along with this, I went through and renamed the objects in our hierarchy
since over the past two weeks they've gotten jumbled. All organized now!