Lasersssssss
Lasers have been implemented in Level 1, Room 5!

There is currently a script on the PlayerController (LaserReset) that
teleports the player to a Vector3 location upon contact with a trigger
with the tag "Laser." Currently, this teleport is fixed to a specific,
in-code location. What needs to be done now is greating checkpoint
triggers that update this Vector3 as the player goes through the levels
(so that a laser three rooms from now doesn't teleport the player all
the way back).