Major Fixes to Raycasting, plus LASERS
Lasers with variable length have been implemented. Lasers travel until
they hit a solid object then stop. An example of such a laser is on the
floor in Level 1 Room 1.

Laser To-Dos:
-- Fix respawning from these new types of lasers (player spawn position
is a bit funky atm).

In order to fix some raycasting issues, I have created a new layer,
"Custom Ignore". This facilitates the player having a larger hitbox
while still being able to shoot downwards. This layer is on the player,
and has been added to all of the boxes/pads in Level 1 (didn't have time
to do Level 2 yet).