A Draft of a Level
Notes:
-Room 3 normally has lasers instead of the floor but because we need to
make the player a child of the platform, and we don’t have the script
for that yet, i added the floor for smoother testing.

-The lasers in the hallway between rooms 3 & 4 were intended to be
animated to move up and down (and the magnet would be pushed under to
prevent them from moving - I’ll look into that tomorrow)

- some magnetized platforms move through the magnetic floor boards,
still don’t understand why?

- room 5 is a very rough draft, just as an idea