Blue Projectiles
*HUGE UPDATE*

Shooting blue now spawns a projectile.
The projectile disappears upon hitting another object (in reality it
despawns after a certain amount of time relative to the distance from
the RayCast hit).
There is also now a delay in color change relative to when the
projectile "hits" the magnetic object.

Also included in this update is a change to the first room of the first
level, experimenting with cube rotation.

To Do:
Add a red projectile