Fixing Old Mistakes
Back when this project was first started, myself and my teammates were
novices with Unity. As such, we did not take advantage of Prefabs to the
extent we should have. For example, the player on each level is its own
separate entity, making edits to the player across all levels tedious.
This update (and the ones to follow) goes through and makes Prefabs
where possible in order to speed up development time down the road.